requirements.txt
numpy
TgCrypto==1.2.0
opencv-python==4.1.1.26
Pyrogram==0.15.1
python-dotenv
